refactor(cli): extract diff/historia + briefing into own modules (2b)
diff/historia → cli_history.py, briefing → cli_briefing.py. cli.py 1291 → 955
LOC (from 2710 at the start of 2b — 65% smaller). What remains inline is the
core/diagnostic surface: stats, doctor, recall-hook, dedupe + the group wiring.
ruff + mypy clean, 833 tests green.
