+++ src/memo/cli.py
@@ -28,20 +28,19 @@ import json
 import os
 import sys
 import time
-from datetime import UTC, datetime
 from pathlib import Path
 from typing import Any
 
 import click
-from rich.panel import Panel
 
 from memo.cli_analytics import analytics_group
 from memo.cli_as_of import as_of_group
 from memo.cli_backend_native import backend_native_group
 from memo.cli_backup import backup_group
+from memo.cli_briefing import briefing
 from memo.cli_capture import capture_stop, ingest, mine_history, resume
 from memo.cli_collaborative import collaborative_group
-from memo.cli_common import _parse_as_of_date, _short, console
+from memo.cli_common import _short, console
 from memo.cli_common import get_memory as _get_memory
 from memo.cli_config import config_group
 from memo.cli_consolidate import consolidate_group
@@ -54,6 +53,7 @@ from memo.cli_eval import eval_group
 from memo.cli_export import export_group
 from memo.cli_feedback import feedback_group
 from memo.cli_graph import graph_group
+from memo.cli_history import diff_cmd, historia_cmd
 from memo.cli_import import import_group
 from memo.cli_links import links_group
 from memo.cli_maintain import maintain_cmd
@@ -131,6 +131,9 @@ cli.add_command(mine_history)
 cli.add_command(ingest)
 cli.add_command(capture_stop)
 cli.add_command(resume)
+cli.add_command(diff_cmd)
+cli.add_command(historia_cmd)
+cli.add_command(briefing)
 cli.add_command(init_cmd)
 cli.add_command(migrate_vault)
 cli.add_command(mcp_command)
@@ -858,150 +861,6 @@ def recall_hook() -> None:
 # ── Recall daemon — persistent socket server for low-latency recall ──────────
 
 
-
-
-
-
-
-
-@cli.command(name="diff")
-@click.option("--from", "from_date", required=True,
-              help="Start date — YYYY-MM-DD or full ISO 8601.")
-@click.option("--to", "to_date", required=False, default=None,
-              help="End date (default: now).")
-@click.option("--json", "as_json", is_flag=True)
-def diff_cmd(from_date: str, to_date: str | None, as_json: bool) -> None:
-    """Diff the corpus between two snapshots.
-
-    Shows added / removed / updated memorias plus a summary line. Useful
-    for "what changed since last Monday" or "what evolved between two
-    releases".
-    """
-    from datetime import datetime as _dt
-
-    from memo.memory import Memory
-    from memo.time_machine import diff as _diff
-
-    to_iso = _dt.now(UTC).isoformat() if to_date is None else _parse_as_of_date(to_date)
-    from_iso = _parse_as_of_date(from_date)
-
-    mem = Memory(Config.from_env())
-    d = _diff(mem, from_ts=from_iso, to_ts=to_iso)
-
-    if as_json:
-        click.echo(json.dumps({
-            "from_ts": d.from_ts.isoformat(),
-            "to_ts": d.to_ts.isoformat(),
-            "added": [{"id": r.id, "title": r.title, "type": r.type} for r in d.added],
-            "removed": [{"id": r.id, "title": r.title, "type": r.type} for r in d.removed],
-            "updated": d.updated,
-        }, ensure_ascii=False, indent=2))
-        return
-
-    console.print(Panel.fit(
-        f"{d.from_ts.date().isoformat()}  →  {d.to_ts.date().isoformat()}\n"
-        f"[bold]{d.summary()}[/bold]",
-        title="corpus diff",
-        border_style="cyan",
-    ))
-    if d.added:
-        console.print(f"\n[green]+ added ({len(d.added)})[/green]")
-        for r in d.added[:20]:
-            console.print(f"  [green]+[/green] [{r.id[:8]}] {r.title}  [dim]({r.type})[/dim]")
-    if d.removed:
-        console.print(f"\n[red]- removed ({len(d.removed)})[/red]")
-        for r in d.removed[:20]:
-            console.print(f"  [red]-[/red] [{r.id[:8]}] {r.title}  [dim]({r.type})[/dim]")
-    if d.updated:
-        console.print(f"\n[yellow]~ updated ({len(d.updated)})[/yellow]")
-        for u in d.updated[:20]:
-            console.print(
-                f"  [yellow]~[/yellow] [{u['id'][:8]}] {u['title']}  "
-                f"[dim](fields: {', '.join(u['changed_fields'])})[/dim]",
-            )
-
-
-@cli.command(name="historia")
-@click.argument("id_or_prefix")
-@click.option("--limit", default=50, type=int, show_default=True,
-              help="Max events to show.")
-@click.option("--json", "as_json", is_flag=True, help="Output raw JSON.")
-def historia_cmd(id_or_prefix: str, limit: int, as_json: bool) -> None:
-    """Show the full edit history for one memoria.
-
-    Displays every save / update / delete event from the audit log,
-    with field-level diffs on each update (title, type, tags, body_hash).
-    Useful for answering "when did I change this?" or reviewing how a
-    decision evolved over time.
-
-    Examples:
-
-      memo historia abc12345
-      memo historia abc12345 --json
-    """
-    from memo.memory import AmbiguousIdError, Memory
-
-    mem = Memory(Config.from_env())
-    try:
-        resolved = mem.resolve_id(id_or_prefix)
-    except AmbiguousIdError as exc:
-        console.print(f"[red]Ambiguous prefix:[/red] {exc}")
-        raise SystemExit(1) from exc
-    if resolved is None:
-        console.print(f"[red]No record found for:[/red] {id_or_prefix!r}")
-        raise SystemExit(1)
-
-    events = mem.history.list_recent(limit=limit, record_id=resolved)
-    events = list(reversed(events))  # chronological order
-
-    if as_json:
-        click.echo(json.dumps(events, ensure_ascii=False, indent=2, default=str))
-        return
-
-    r = mem.get(resolved)
-    title_str = f"{r.title}" if r else resolved[:8]
-    console.print(Panel.fit(
-        f"[bold]{title_str}[/bold]  [dim]{resolved[:8]}[/dim]",
-        title="historia",
-        border_style="cyan",
-    ))
-
-    if not events:
-        console.print("  [dim](no events in audit log)[/dim]")
-        return
-
-    _OP_STYLE = {"save": "green", "update": "yellow", "delete": "red"}
-
-    for ev in events:
-        op = ev.get("op", "?")
-        ts = ev.get("ts", "")
-        style = _OP_STYLE.get(op, "white")
-        ts_short = ts[:16].replace("T", " ") if ts else "?"
-        console.print(f"\n  [{style}]{op.upper():6s}[/{style}]  [dim]{ts_short}[/dim]")
-
-        delta = ev.get("delta")
-        if not delta:
-            continue
-        for field, pair in delta.items():
-            if not isinstance(pair, list) or len(pair) != 2:
-                continue
-            old_v, new_v = pair
-            if field == "tags":
-                old_s = ", ".join(old_v) if isinstance(old_v, list) else str(old_v)
-                new_s = ", ".join(new_v) if isinstance(new_v, list) else str(new_v)
-            elif field == "body_hash":
-                old_s, new_s = str(old_v)[:12], str(new_v)[:12]
-            else:
-                old_s, new_s = str(old_v), str(new_v)
-            console.print(
-                f"           [dim]{field}:[/dim]  "
-                f"[red]{old_s}[/red]  →  [green]{new_s}[/green]"
-            )
-
-    last_ts = events[-1].get("ts", "")
-    console.print(f"\n  [dim]{len(events)} event(s) · last: {last_ts[:16].replace('T', ' ')}[/dim]")
-
-
 @cli.command(name="dedupe")
 @click.option("--threshold", type=float, default=0.92,
               help="Cosine threshold for near-duplicate clustering (default: 0.92)")
@@ -1089,199 +948,6 @@ def dedupe_cmd(
             f"{result.merged_id[:8] if result.merged_id else 'n/a'}  "
             f"archived={len(result.archived_ids)}"
         )
-
-
-@cli.command(name="briefing")
-def briefing() -> None:
-    """SessionStart hook — rich context panel.
-
-    Emits a `hookSpecificOutput` JSON with `additionalContext` markdown
-    containing:
-      - Last session for the current project (crash recovery)
-      - Open loops: recently updated memories (in-flight decisions)
-      - Memory of the day: one memory picked deterministically by date
-      - Quick interaction guide
-
-    All errors swallowed — a failed briefing is worse than no briefing
-    only if it blocks the session. Exit 0 + `{}` on any failure.
-
-    Env vars:
-      MEMO_BRIEFING_DISABLE        — set to "1" to skip entirely
-      MEMO_BRIEFING_LOOPS_N        — open-loop count to show (default 5)
-      MEMO_BRIEFING_LOOPS_DAYS     — how recent counts as "open" (default 7)
-      MEMO_BRIEFING_DEBUG          — print errors to stderr
-    """
-    import hashlib as _hashlib
-    import json as _json
-    import os as _os
-    import sys as _sys
-    from datetime import timedelta
-
-    debug = _os.environ.get("MEMO_BRIEFING_DEBUG") == "1"
-
-    def _bail(reason: str = "") -> None:
-        if reason and debug:
-            print(f"# memo briefing: {reason}", file=_sys.stderr)
-        print("{}")
-        _sys.exit(0)
-
-    if _os.environ.get("MEMO_BRIEFING_DISABLE") == "1":
-        _bail("disabled")
-        return
-
-    try:
-        cfg = Config.from_env()
-        from memo.memory import Memory
-        mem = Memory(cfg)
-    except Exception as exc:
-        _bail(f"Memory init failed: {exc}")
-        return
-
-    loops_n = max(1, int(_os.environ.get("MEMO_BRIEFING_LOOPS_N", "5") or 5))
-    loops_days = max(1, int(_os.environ.get("MEMO_BRIEFING_LOOPS_DAYS", "7") or 7))
-
-    lines: list[str] = []
-
-    # ── 1. Last session for this project ──────────────────────────────────
-    try:
-        from pathlib import Path as _Path
-
-        from memo.session import format_relative, list_sessions
-
-        cur_cwd = str(_Path(_os.getcwd()).resolve())
-        all_sessions = list_sessions(cfg.state_dir, limit=20)
-        same_proj = [r for r in all_sessions if (r.get("cwd") or "") == cur_cwd]
-        if same_proj:
-            top = same_proj[0]
-            sid = top.get("session_id") or ""
-            when = format_relative(top.get("updated"))
-            summary = (
-                top.get("summary") or top.get("last_user_msg") or "—"
-            ).replace("\n", " ")[:120]
-            lines.append("## El Briefing")
-            lines.append("")
-            lines.append(f"**Última sesión en este proyecto** ({when}): {summary}")
-            lines.append(f"`claude --resume {sid}`")
-            lines.append("")
-    except Exception as exc:
-        if debug:
-            print(f"# memo briefing: session lookup failed: {exc}", file=_sys.stderr)
-        if not lines:
-            lines.append("## El Briefing")
-            lines.append("")
-
-    # ── 1b. Unified consciousness (Synapse) ───────────────────────────────
-    # Pulls present_state (memflow handoffs/focus) + reality_conflicts from
-    # `synapse packet`. No-op when synapse is not installed or unreachable —
-    # the rest of the briefing is unaffected (graceful, opt-in boundary).
-    if _os.environ.get("MEMO_BRIEFING_SYNAPSE_DISABLE") != "1":
-        try:
-            from memo.briefing import synapse_briefing_lines
-
-            try:
-                cur_cwd_str = _os.getcwd()
-            except Exception:
-                cur_cwd_str = ""
-            syn_lines = synapse_briefing_lines(cur_cwd_str)
-            if syn_lines:
-                lines.extend(syn_lines)
-        except Exception as exc:
-            if debug:
-                print(f"# memo briefing: synapse lookup failed: {exc}", file=_sys.stderr)
-
-    # ── 2. Open loops: recently updated memories ──────────────────────────
-    try:
-        cutoff = (datetime.now(tz=UTC) - timedelta(days=loops_days)).isoformat()
-        all_recent = mem.store.list_recent(limit=loops_n * 4)
-        open_loops = [
-            r for r in all_recent
-            if (r.get("updated") or "") >= cutoff
-        ][:loops_n]
-
-        if open_loops:
-            lines.append(f"### Loops abiertos (últimos {loops_days} días)")
-            lines.append("")
-            for i, r in enumerate(open_loops, start=1):
-                tags = r.get("tags") or []
-                if isinstance(tags, str):
-                    import json as _j
-                    try:
-                        tags = _j.loads(tags)
-                    except Exception:
-                        tags = []
-                tag_str = ", ".join(str(t) for t in tags[:3]) if tags else ""
-                title = r.get("title") or "—"
-                type_ = r.get("type") or "note"
-                id_short = (r.get("id") or "")[:8]
-                updated = r.get("updated") or ""
-                try:
-                    dt = datetime.fromisoformat(updated)
-                    if dt.tzinfo is None:
-                        dt = dt.replace(tzinfo=UTC)
-                    delta = datetime.now(tz=UTC) - dt
-                    days_ago = delta.days
-                    age = f"hace {days_ago}d" if days_ago > 0 else "hoy"
-                except Exception:
-                    age = updated[:10]
-                lines.append(
-                    f"{i}. `{id_short}` **{type_}** · {title}"
-                    + (f" — {age}" if age else "")
-                    + (f" [{tag_str}]" if tag_str else "")
-                )
-            lines.append("")
-    except Exception as exc:
-        if debug:
-            print(f"# memo briefing: open-loops failed: {exc}", file=_sys.stderr)
-
-    # ── 3. Memory of the day (date-seeded, biased to least-recent) ────────
-    try:
-        # Use today's date as seed so the pick is stable within a day but
-        # rotates daily. Favour memories whose `updated` is oldest (least
-        # recently revisited) so the corpus gets covered over time.
-        today_str = datetime.now(tz=UTC).strftime("%Y-%m-%d")
-        all_ids_rows = mem.store.list_recent(limit=500)
-        if all_ids_rows:
-            # Sort oldest-updated first so the seed picks from the back of
-            # the corpus on average.
-            sorted_rows = sorted(all_ids_rows, key=lambda r: r.get("updated") or "")
-            seed_int = int(_hashlib.sha256(today_str.encode()).hexdigest(), 16)
-            pick_row = sorted_rows[seed_int % len(sorted_rows)]
-            pick_id = pick_row.get("id") or ""
-            pick_rec = mem.get(pick_id) if pick_id else None
-            if pick_rec:
-                body_preview = (pick_rec.body or "").strip()[:200].replace("\n", " ")
-                tags = pick_rec.tags or []
-                tag_str = ", ".join(str(t) for t in tags[:4]) if tags else ""
-                lines.append("### Memoria del día")
-                lines.append("")
-                lines.append(
-                    f"`{pick_rec.id[:8]}` **{pick_rec.type}** · {pick_rec.title}"
-                    + (f" [{tag_str}]" if tag_str else "")
-                )
-                if body_preview:
-                    lines.append(f"> {body_preview}{'…' if len(pick_rec.body or '') > 200 else ''}")
-                lines.append("")
-    except Exception as exc:
-        if debug:
-            print(f"# memo briefing: memory-of-day failed: {exc}", file=_sys.stderr)
-
-    # ── 4. Interaction guide ──────────────────────────────────────────────
-    lines.append(
-        "_Para continuar: `dame el loop N` (retoma por número) · "
-        "`/memo get <id>` · `/memo ask <pregunta>`_"
-    )
-
-    if not any(ln for ln in lines if ln and not ln.startswith("#") and not ln.startswith("_")):
-        _bail("nothing to show")
-        return
-
-    output = {
-        "hookSpecificOutput": {
-            "hookEventName": "SessionStart",
-            "additionalContext": "\n".join(lines),
-        }
-    }
-    print(_json.dumps(output, ensure_ascii=False))
 
 
 def main() -> None:
