+++ src/memo/cli_briefing.py
@@ -1,0 +1,207 @@
+"""`memo briefing` — El Briefing (open loops + memory of the day) for SessionStart.
+
+Extracted from cli.py (2b god-module decomposition). Registered via
+`cli.add_command(briefing)`.
+"""
+
+from __future__ import annotations
+
+from datetime import UTC, datetime
+
+import click
+
+from memo.config import Config
+
+
+@click.command(name="briefing")
+def briefing() -> None:
+    """SessionStart hook — rich context panel.
+
+    Emits a `hookSpecificOutput` JSON with `additionalContext` markdown
+    containing:
+      - Last session for the current project (crash recovery)
+      - Open loops: recently updated memories (in-flight decisions)
+      - Memory of the day: one memory picked deterministically by date
+      - Quick interaction guide
+
+    All errors swallowed — a failed briefing is worse than no briefing
+    only if it blocks the session. Exit 0 + `{}` on any failure.
+
+    Env vars:
+      MEMO_BRIEFING_DISABLE        — set to "1" to skip entirely
+      MEMO_BRIEFING_LOOPS_N        — open-loop count to show (default 5)
+      MEMO_BRIEFING_LOOPS_DAYS     — how recent counts as "open" (default 7)
+      MEMO_BRIEFING_DEBUG          — print errors to stderr
+    """
+    import hashlib as _hashlib
+    import json as _json
+    import os as _os
+    import sys as _sys
+    from datetime import timedelta
+
+    debug = _os.environ.get("MEMO_BRIEFING_DEBUG") == "1"
+
+    def _bail(reason: str = "") -> None:
+        if reason and debug:
+            print(f"# memo briefing: {reason}", file=_sys.stderr)
+        print("{}")
+        _sys.exit(0)
+
+    if _os.environ.get("MEMO_BRIEFING_DISABLE") == "1":
+        _bail("disabled")
+        return
+
+    try:
+        cfg = Config.from_env()
+        from memo.memory import Memory
+        mem = Memory(cfg)
+    except Exception as exc:
+        _bail(f"Memory init failed: {exc}")
+        return
+
+    loops_n = max(1, int(_os.environ.get("MEMO_BRIEFING_LOOPS_N", "5") or 5))
+    loops_days = max(1, int(_os.environ.get("MEMO_BRIEFING_LOOPS_DAYS", "7") or 7))
+
+    lines: list[str] = []
+
+    # ── 1. Last session for this project ──────────────────────────────────
+    try:
+        from pathlib import Path as _Path
+
+        from memo.session import format_relative, list_sessions
+
+        cur_cwd = str(_Path(_os.getcwd()).resolve())
+        all_sessions = list_sessions(cfg.state_dir, limit=20)
+        same_proj = [r for r in all_sessions if (r.get("cwd") or "") == cur_cwd]
+        if same_proj:
+            top = same_proj[0]
+            sid = top.get("session_id") or ""
+            when = format_relative(top.get("updated"))
+            summary = (
+                top.get("summary") or top.get("last_user_msg") or "—"
+            ).replace("\n", " ")[:120]
+            lines.append("## El Briefing")
+            lines.append("")
+            lines.append(f"**Última sesión en este proyecto** ({when}): {summary}")
+            lines.append(f"`claude --resume {sid}`")
+            lines.append("")
+    except Exception as exc:
+        if debug:
+            print(f"# memo briefing: session lookup failed: {exc}", file=_sys.stderr)
+        if not lines:
+            lines.append("## El Briefing")
+            lines.append("")
+
+    # ── 1b. Unified consciousness (Synapse) ───────────────────────────────
+    # Pulls present_state (memflow handoffs/focus) + reality_conflicts from
+    # `synapse packet`. No-op when synapse is not installed or unreachable —
+    # the rest of the briefing is unaffected (graceful, opt-in boundary).
+    if _os.environ.get("MEMO_BRIEFING_SYNAPSE_DISABLE") != "1":
+        try:
+            from memo.briefing import synapse_briefing_lines
+
+            try:
+                cur_cwd_str = _os.getcwd()
+            except Exception:
+                cur_cwd_str = ""
+            syn_lines = synapse_briefing_lines(cur_cwd_str)
+            if syn_lines:
+                lines.extend(syn_lines)
+        except Exception as exc:
+            if debug:
+                print(f"# memo briefing: synapse lookup failed: {exc}", file=_sys.stderr)
+
+    # ── 2. Open loops: recently updated memories ──────────────────────────
+    try:
+        cutoff = (datetime.now(tz=UTC) - timedelta(days=loops_days)).isoformat()
+        all_recent = mem.store.list_recent(limit=loops_n * 4)
+        open_loops = [
+            r for r in all_recent
+            if (r.get("updated") or "") >= cutoff
+        ][:loops_n]
+
+        if open_loops:
+            lines.append(f"### Loops abiertos (últimos {loops_days} días)")
+            lines.append("")
+            for i, r in enumerate(open_loops, start=1):
+                tags = r.get("tags") or []
+                if isinstance(tags, str):
+                    import json as _j
+                    try:
+                        tags = _j.loads(tags)
+                    except Exception:
+                        tags = []
+                tag_str = ", ".join(str(t) for t in tags[:3]) if tags else ""
+                title = r.get("title") or "—"
+                type_ = r.get("type") or "note"
+                id_short = (r.get("id") or "")[:8]
+                updated = r.get("updated") or ""
+                try:
+                    dt = datetime.fromisoformat(updated)
+                    if dt.tzinfo is None:
+                        dt = dt.replace(tzinfo=UTC)
+                    delta = datetime.now(tz=UTC) - dt
+                    days_ago = delta.days
+                    age = f"hace {days_ago}d" if days_ago > 0 else "hoy"
+                except Exception:
+                    age = updated[:10]
+                lines.append(
+                    f"{i}. `{id_short}` **{type_}** · {title}"
+                    + (f" — {age}" if age else "")
+                    + (f" [{tag_str}]" if tag_str else "")
+                )
+            lines.append("")
+    except Exception as exc:
+        if debug:
+            print(f"# memo briefing: open-loops failed: {exc}", file=_sys.stderr)
+
+    # ── 3. Memory of the day (date-seeded, biased to least-recent) ────────
+    try:
+        # Use today's date as seed so the pick is stable within a day but
+        # rotates daily. Favour memories whose `updated` is oldest (least
+        # recently revisited) so the corpus gets covered over time.
+        today_str = datetime.now(tz=UTC).strftime("%Y-%m-%d")
+        all_ids_rows = mem.store.list_recent(limit=500)
+        if all_ids_rows:
+            # Sort oldest-updated first so the seed picks from the back of
+            # the corpus on average.
+            sorted_rows = sorted(all_ids_rows, key=lambda r: r.get("updated") or "")
+            seed_int = int(_hashlib.sha256(today_str.encode()).hexdigest(), 16)
+            pick_row = sorted_rows[seed_int % len(sorted_rows)]
+            pick_id = pick_row.get("id") or ""
+            pick_rec = mem.get(pick_id) if pick_id else None
+            if pick_rec:
+                body_preview = (pick_rec.body or "").strip()[:200].replace("\n", " ")
+                tags = pick_rec.tags or []
+                tag_str = ", ".join(str(t) for t in tags[:4]) if tags else ""
+                lines.append("### Memoria del día")
+                lines.append("")
+                lines.append(
+                    f"`{pick_rec.id[:8]}` **{pick_rec.type}** · {pick_rec.title}"
+                    + (f" [{tag_str}]" if tag_str else "")
+                )
+                if body_preview:
+                    lines.append(f"> {body_preview}{'…' if len(pick_rec.body or '') > 200 else ''}")
+                lines.append("")
+    except Exception as exc:
+        if debug:
+            print(f"# memo briefing: memory-of-day failed: {exc}", file=_sys.stderr)
+
+    # ── 4. Interaction guide ──────────────────────────────────────────────
+    lines.append(
+        "_Para continuar: `dame el loop N` (retoma por número) · "
+        "`/memo get <id>` · `/memo ask <pregunta>`_"
+    )
+
+    if not any(ln for ln in lines if ln and not ln.startswith("#") and not ln.startswith("_")):
+        _bail("nothing to show")
+        return
+
+    output = {
+        "hookSpecificOutput": {
+            "hookEventName": "SessionStart",
+            "additionalContext": "\n".join(lines),
+        }
+    }
+    print(_json.dumps(output, ensure_ascii=False))
+
